Atualização lista tres
